NIFI-8025 - Removed some repetitive code. Moved some logout code up to the AccessResource class.
